Remove AccountAdmin components
